requirements.txt
discord.py
flask
requests
beautifulsoup4
feedparser
